회원 탈퇴, 수정, 정보보기 시 권한 검증

+++ src/main/java/com/example/prj1be/controller/MemberController.java
@@ -71,9 +71,15 @@ public class MemberController {
 
     // 회원 목록 상세 보기 컨트롤러
     @GetMapping
-    public ResponseEntity<Member> view(String id) {
-        // TODO : 로그인 했는 지? => 안 했으면 401
-        // TODO : 자기 정보인지? => 아니면 403
+    public ResponseEntity<Member> view(String id,
+                                       @SessionAttribute(value = "login", required = false) Member login) {
+        if (login == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+
+        if (!service.hasAccess(id, login)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
 
         Member member = service.getMember(id);
 
@@ -82,9 +88,15 @@ public class MemberController {
 
     // 회원 목록 상세보기 - 회원 탈퇴 기능
     @DeleteMapping
-    public ResponseEntity delete(String id) {
-        // TODO : 로그인 했는 지? => 안 했으면 401
-        // TODO : 자기 정보인지? => 아니면 403
+    public ResponseEntity delete(String id,
+                                 @SessionAttribute(value = "login", required = false) Member login) {
+        if (login == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build(); // 401
+        }
+
+        if (!service.hasAccess(id, login)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build(); // 403
+        }
 
         if (service.deleteMember(id)) {
             return ResponseEntity.ok().build();
@@ -94,8 +106,15 @@ public class MemberController {
 
     // 회원목록상세 - 회원수정
     @PutMapping("edit")
-    public ResponseEntity edit(@RequestBody Member member) {
-        // TODO : 로그인 했는지? 자기정보인지?
+    public ResponseEntity edit(@RequestBody Member member,
+                               @SessionAttribute(value = "login",required = false)Member login) {
+
+        if (login == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+        if (!service.hasAccess(member.getId(), login)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
 
         if (service.update(member)) {
             return ResponseEntity.ok().build();
